requirements.txt
Django==3.2
django-environ==0.4.5
djangorestframework==3.12.4
django-debug-toolbar==3.2.1
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0
Pillow==8.2.0